# src/main/java/org/example/dto/responses/TaskResponse.java
package org.example.dto.responses;

import lombok.Data;
import org.example.entities.CommentEntity;
import org.example.entities.TaskEntity;

import java.time.LocalDateTime;
import java.util.List;

@Data
public class TaskResponse {
    private final long id;
    private final String title;
    private final String description;
    private final LocalDateTime creationTime;
    private final LocalDateTime lastUpdateTime;
    private final LocalDateTime dueTime;
    private final List<CommentEntity> comments;
    private final String reporter;
    private final String assignee;

    public TaskResponse(TaskEntity taskEntity) {
        this.id = taskEntity.getId();
        this.title = taskEntity.getTitle();
        this.description = taskEntity.getDescription();
        this.creationTime = taskEntity.getCreationTime();
        this.lastUpdateTime = taskEntity.getLastUpdateTime();
        this.dueTime = taskEntity.getDueTime();
        this.comments = taskEntity.getComments();
        this.reporter = taskEntity.getReporter().getName();
        this.assignee = taskEntity.getAssignee().getName();
    }
}